Move the network default conf in his settings.rs
