+++ massa-network/src/settings.rs
@@ -60,6 +60,7 @@ pub struct NetworkSettings {
 #[cfg(test)]
 mod tests {
     use crate::NetworkSettings;
+    use massa_models::constants::*;
     use massa_time::MassaTime;
     use std::net::{IpAddr, Ipv4Addr, SocketAddr};
 
@@ -89,6 +90,50 @@ mod tests {
                 initial_peers_file: std::path::PathBuf::new(),
                 peer_list_send_timeout: MassaTime::from(500),
                 max_in_connection_overflow: 2,
+            }
+        }
+    }
+
+    impl NetworkSettings {
+        pub fn scenarios_default(port: u16, peers_file: &std::path::Path) -> Self {
+            // Init the serialization context with a default,
+            // can be overwritten with a more specific one in the test.
+            massa_models::init_serialization_context(massa_models::SerializationContext {
+                max_advertise_length: 128,
+                max_bootstrap_blocks: 100,
+                max_bootstrap_cliques: 100,
+                max_bootstrap_deps: 100,
+                max_bootstrap_children: 100,
+                max_ask_blocks_per_message: 10,
+                endorsement_count: 8,
+                ..massa_models::SerializationContext::default()
+            });
+            Self {
+                bind: format!("0.0.0.0:{}", port).parse().unwrap(),
+                routable_ip: Some(BASE_NETWORK_CONTROLLER_IP),
+                protocol_port: port,
+                connect_timeout: MassaTime::from(3000),
+                peers_file: peers_file.to_path_buf(),
+                target_out_nonbootstrap_connections: 10,
+                wakeup_interval: MassaTime::from(3000),
+                target_bootstrap_connections: 0,
+                max_out_bootstrap_connection_attempts: 1,
+                max_in_nonbootstrap_connections: 100,
+                max_in_connections_per_ip: 100,
+                max_out_nonbootstrap_connection_attempts: 100,
+                max_idle_peers: 100,
+                max_banned_peers: 100,
+                peers_file_dump_interval: MassaTime::from(30000),
+                message_timeout: MassaTime::from(5000u64),
+                ask_peer_list_interval: MassaTime::from(50000u64),
+                private_key_file: crate::tests::tools::get_temp_private_key_file()
+                    .path()
+                    .to_path_buf(),
+                max_send_wait: MassaTime::from(100),
+                ban_timeout: MassaTime::from(100_000_000),
+                initial_peers_file: peers_file.to_path_buf(),
+                peer_list_send_timeout: MassaTime::from(50),
+                max_in_connection_overflow: 10,
             }
         }
     }
