+++ massa-network/src/tests/mock_establisher.rs
@@ -13,12 +13,11 @@ const MAX_DUPLEX_BUFFER_SIZE: usize = 1024;
 pub type ReadHalf = tokio::io::ReadHalf<DuplexStream>;
 pub type WriteHalf = tokio::io::WriteHalf<DuplexStream>;
 
-pub type Receiver = mpsc::Receiver<(SocketAddr, oneshot::Sender<(ReadHalf, WriteHalf)>)>;
-pub type Sender = mpsc::Sender<(SocketAddr, oneshot::Sender<(ReadHalf, WriteHalf)>)>;
+type AddrSender = (SocketAddr, oneshot::Sender<(ReadHalf, WriteHalf)>);
 
 pub fn new() -> (MockEstablisher, MockEstablisherInterface) {
     let (connection_listener_tx, connection_listener_rx) =
-        mpsc::channel::<(SocketAddr, oneshot::Sender<(ReadHalf, WriteHalf)>)>(CHANNEL_SIZE);
+        mpsc::channel::<AddrSender>(CHANNEL_SIZE);
 
     let (connection_connector_tx, connection_connector_rx) =
         mpsc::channel::<(ReadHalf, WriteHalf, SocketAddr, oneshot::Sender<bool>)>(CHANNEL_SIZE);
@@ -37,7 +36,7 @@ pub fn new() -> (MockEstablisher, MockEstablisherInterface) {
 
 #[derive(Debug)]
 pub struct MockListener {
-    connection_listener_rx: Receiver, // (controller, mock)
+    connection_listener_rx: mpsc::Receiver<AddrSender>, // (controller, mock)
 }
 
 impl MockListener {
@@ -112,7 +111,7 @@ impl MockConnector {
 
 #[derive(Debug)]
 pub struct MockEstablisher {
-    connection_listener_rx: Option<Receiver>,
+    connection_listener_rx: Option<mpsc::Receiver<AddrSender>>,
     connection_connector_tx: mpsc::Sender<(ReadHalf, WriteHalf, SocketAddr, oneshot::Sender<bool>)>,
 }
 
@@ -140,7 +139,7 @@ impl MockEstablisher {
 }
 
 pub struct MockEstablisherInterface {
-    connection_listener_tx: Option<Sender>,
+    connection_listener_tx: Option<mpsc::Sender<AddrSender>>,
     connection_connector_rx:
         mpsc::Receiver<(ReadHalf, WriteHalf, SocketAddr, oneshot::Sender<bool>)>,
 }
