+++ massa-network/src/tests/mod.rs
@@ -2,4 +2,5 @@
 
 pub mod mock_establisher;
 mod scenarios;
+mod test_peer_info_database;
 pub mod tools;
