+++ massa-network/src/tests/tools.rs
@@ -22,16 +22,13 @@ use massa_time::MassaTime;
 use std::str::FromStr;
 use std::{
     future::Future,
-    net::{IpAddr, Ipv4Addr, SocketAddr},
-    path::Path,
+    net::{IpAddr, SocketAddr},
     time::Duration,
 };
 use tempfile::NamedTempFile;
 use tokio::time::sleep;
 use tokio::{sync::oneshot, task::JoinHandle, time::timeout};
 use tracing::trace;
-
-pub const BASE_NETWORK_CONTROLLER_IP: IpAddr = IpAddr::V4(Ipv4Addr::new(169, 202, 0, 10));
 
 pub fn get_dummy_block_id(s: &str) -> BlockId {
     BlockId(Hash::compute_from(s.as_bytes()))
@@ -50,62 +47,8 @@ pub fn generate_peers_file(peer_vec: &[PeerInfo]) -> NamedTempFile {
     peers_file_named
 }
 
-fn get_temp_private_key_file() -> NamedTempFile {
+pub fn get_temp_private_key_file() -> NamedTempFile {
     NamedTempFile::new().expect("cannot create temp file")
-}
-
-/// create a NetworkConfig with typical values
-pub fn create_network_config(
-    network_controller_port: u16,
-    peers_file_path: &Path,
-) -> NetworkSettings {
-    // Init the serialization context with a default,
-    // can be overwritten with a more specific one in the test.
-    massa_models::init_serialization_context(massa_models::SerializationContext {
-        max_operations_per_block: 1024,
-        thread_count: 2,
-        max_advertise_length: 128,
-        max_message_size: 3 * 1024 * 1024,
-        max_block_size: 3 * 1024 * 1024,
-        max_bootstrap_blocks: 100,
-        max_bootstrap_cliques: 100,
-        max_bootstrap_deps: 100,
-        max_bootstrap_children: 100,
-        max_ask_blocks_per_message: 10,
-        max_operations_per_message: 1024,
-        max_endorsements_per_message: 1024,
-        max_bootstrap_message_size: 100000000,
-        max_bootstrap_pos_entries: 1000,
-        max_bootstrap_pos_cycles: 5,
-        endorsement_count: 8,
-    });
-
-    NetworkSettings {
-        bind: format!("0.0.0.0:{}", network_controller_port)
-            .parse()
-            .unwrap(),
-        routable_ip: Some(BASE_NETWORK_CONTROLLER_IP),
-        protocol_port: network_controller_port,
-        connect_timeout: MassaTime::from(3000),
-        peers_file: peers_file_path.to_path_buf(),
-        target_out_nonbootstrap_connections: 10,
-        wakeup_interval: MassaTime::from(3000),
-        target_bootstrap_connections: 0,
-        max_out_bootstrap_connection_attempts: 1,
-        max_in_nonbootstrap_connections: 100,
-        max_in_connections_per_ip: 100,
-        max_out_nonbootstrap_connection_attempts: 100,
-        max_idle_peers: 100,
-        max_banned_peers: 100,
-        peers_file_dump_interval: MassaTime::from(30000),
-        message_timeout: MassaTime::from(5000u64),
-        ask_peer_list_interval: MassaTime::from(50000u64),
-        private_key_file: get_temp_private_key_file().path().to_path_buf(),
-        max_send_wait: MassaTime::from(100),
-        ban_timeout: MassaTime::from(100_000_000),
-        initial_peers_file: peers_file_path.to_path_buf(),
-        ..Default::default()
-    }
 }
 
 /// Establish a full alive connection to the controller
